Added resours base Authorization

+++ Timesheets/Controllers/TimesheetEntriesController.cs
@@ -3,12 +3,14 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Timesheets.Data;
 using Timesheets.Models;
+using Timesheets.Security;
 
 namespace Timesheets.Controllers
 {
@@ -16,21 +18,51 @@ namespace Timesheets.Controllers
     {
         private readonly ApplicationDbContext _context;
         private UserManager<MyUser> UserManager;
+        private readonly IAuthorizationService _authorizationService;
 
-        public TimesheetEntriesController(ApplicationDbContext context,UserManager<MyUser> userManager)
+        public TimesheetEntriesController(ApplicationDbContext context,UserManager<MyUser> userManager
+            , IAuthorizationService authorizationService)
         {
             _context = context;
             UserManager = userManager;
+            _authorizationService = authorizationService;
         }
 
         // GET: TimesheetEntries
+        [Authorize(Roles = "Admin , Manager , Employee")]
         public async Task<IActionResult> Index()
         {
-            
-            return View(await _context.TimesheetEntries.Include( p => p.RelatedProject).Include(u=>u.RelatedUser).ToListAsync());
+            //var usrs = UserManager.Users.ToList();
+            //var useer = _context.Users.ToList();
+            MyUser user = await UserManager.GetUserAsync(HttpContext.User);
+           // var alluser = await UserManager.GetUsers
+            var allTimesheets = await _context.TimesheetEntries.Include(p => p.RelatedProject).Include(u => u.RelatedUser).ToListAsync();
+
+            if (User.IsInRole("Admin"))
+            {       
+
+                return View(allTimesheets);
+            }
+            else if (User.IsInRole("Manager"))
+            {
+                 var certainTimesheets = from timesheet in allTimesheets
+                                        where timesheet.RelatedUser.Department.Id == user.Department.Id
+                                        select timesheet;
+                return View(certainTimesheets);
+            }
+            else if (User.IsInRole("Employee")) {
+
+                  var certainTimesheets = from timesheet in allTimesheets
+                                        where timesheet.RelatedUser.Id == user.Id
+                                        select timesheet;
+                return View(certainTimesheets);
+            }
+            else return View(nameof(Error), "Please log in");
         }
 
         // GET: TimesheetEntries/Details/5
+        [Authorize(Roles = "Admin,Manager,Employee")]
+      
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
@@ -40,18 +72,25 @@ namespace Timesheets.Controllers
 
             var timesheetEntry = await _context.TimesheetEntries.Include(p=>p.RelatedProject).Include(u=>u.RelatedUser)
                 .FirstOrDefaultAsync(m => m.Id == id);
-           // var tims = await _context.TimesheetEntries.Include(p => p.RelatedProject).Include(u => u.RelatedUser).ToListAsync();
-          //  var actual = from tim in tims where tim.RelatedProject.Id == 1 select tim;
             if (timesheetEntry == null)
             {
                 return NotFound();
             }
+            var result=await _authorizationService.AuthorizeAsync(User, timesheetEntry, Operations.Read);
+            if (result.Succeeded){
+                return View(timesheetEntry);
+            }
+            else{
+                return View(nameof(Error), "Access Denied");
+            }
+            
             
 
-            return View(timesheetEntry);
+            
         }
 
         // GET: TimesheetEntries/Create
+        [Authorize(Roles = "Admin,Manager,Employee")]
         public IActionResult Create()
         {
             var projects = new List<SelectListItem>();
@@ -69,6 +108,7 @@ namespace Timesheets.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Admin,Manager,Employee")]
         public async Task<IActionResult> Create( TimesheetEntryViewModel timesheetEntry)
         {
             
@@ -95,9 +135,11 @@ namespace Timesheets.Controllers
                         DateCreated = timesheetEntry.DateCreated,
                         HoursWorked = timesheetEntry.HoursWorked
                     };
-                    _context.Add(actualTimesheetEntry);
-                    await _context.SaveChangesAsync();
-                    return RedirectToAction(nameof(Index));
+                    
+                        _context.Add(actualTimesheetEntry);
+                        await _context.SaveChangesAsync();
+                        return RedirectToAction(nameof(Index));
+                    
                 }
                 else {
                     var ErrorMsg = "Timesheet for Project: " + actualProject.Name +
@@ -109,6 +151,9 @@ namespace Timesheets.Controllers
         }
 
         // GET: TimesheetEntries/Edit/5
+       
+
+        [Authorize(Roles = "Admin,Manager")]
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -135,8 +180,14 @@ namespace Timesheets.Controllers
                 HoursWorked=timesheetEntry.HoursWorked
             };
 
-
-            return View(timesheetViewModel);
+            var result = await _authorizationService.AuthorizeAsync(User, timesheetEntry, Operations.Update);
+            if (result.Succeeded)
+            {
+                return View(timesheetViewModel);
+            }
+            else {
+                return View(nameof(Error), "Access Denied");
+            }
         }
 
         // POST: TimesheetEntries/Edit/5
@@ -144,6 +195,8 @@ namespace Timesheets.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        
+        [Authorize(Roles = "Admin,Manager")]
         public async Task<IActionResult> Edit(int id,  TimesheetEntryViewModel timesheetEntry)
         {
             var actualProject = _context.Projects.Find(timesheetEntry.RelatedProject);
@@ -155,24 +208,21 @@ namespace Timesheets.Controllers
                 try
                 {
                     TimesheetEntry actualTimesheet = _context.TimesheetEntries.Find(id);
-                    var possibleTimesheet = from tim in allTimesheets
-                                            where tim.RelatedProject.Id == actualProject.Id &&
-                                            tim.RelatedUser.Id == actualTimesheet.RelatedUser.Id &&
-                                            tim.DateCreated.Date == timesheetEntry.DateCreated.Date
-                                            select tim;
-                    if (possibleTimesheet.Count() == 0)
-                    {
+                    
+                   
                         actualTimesheet.RelatedProject = actualProject;
                         actualTimesheet.HoursWorked = timesheetEntry.HoursWorked;
                         actualTimesheet.DateCreated = timesheetEntry.DateCreated;
+                    var result = await _authorizationService.AuthorizeAsync(User, actualTimesheet, Operations.Update);
+                    if (result.Succeeded)
+                    {
                         _context.Update(actualTimesheet);
                         await _context.SaveChangesAsync();
                     }
-                    else {
-                        var ErrorMsg = "Timesheet for Project: " + actualProject.Name +
-                        " with Date: " + timesheetEntry.DateCreated.Date + " already exists";
-                        return View(nameof(Error), ErrorMsg);
+                    else{
+                        return View(nameof(Error), "Access Denied");
                     }
+                   
                 }
                 catch (DbUpdateConcurrencyException)
                 {
@@ -191,6 +241,9 @@ namespace Timesheets.Controllers
         }
 
         // GET: TimesheetEntries/Delete/5
+        [Authorize(Roles = "Admin,Manager")]
+       
+
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
@@ -204,19 +257,35 @@ namespace Timesheets.Controllers
             {
                 return NotFound();
             }
-
-            return View(timesheetEntry);
+            var result = await _authorizationService.AuthorizeAsync(User, timesheetEntry, Operations.Delete);
+            if (result.Succeeded)
+            {
+                return View(timesheetEntry);
+            }
+            else {
+                return View(nameof(Error), "Access Denied");
+            }
         }
 
         // POST: TimesheetEntries/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Admin,Manager")]
+       
+
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var timesheetEntry = await _context.TimesheetEntries.FindAsync(id);
-            _context.TimesheetEntries.Remove(timesheetEntry);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            var result = await _authorizationService.AuthorizeAsync(User, timesheetEntry, Operations.Delete);
+            if (result.Succeeded)
+            {
+                _context.TimesheetEntries.Remove(timesheetEntry);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            else {
+                return View(nameof(Error), "AccessDenied");
+            }
         }
 
         private bool TimesheetEntryExists(int id)
